requirements.txt
ollama
gymnasium[accept-rom-license, box2d]
openai
pyyaml
imageio